requirements.txt
absl-py==2.1.0
accelerate==1.10.1
aiofiles==24.1.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2024.8.30
chex==0.1.89
click==8.1.7
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diffrax==0.6.0
diffusers==0.35.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.16
equinox==0.11.6
etils==1.9.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
flatbuffers==24.3.25
flax==0.10.5
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanize==4.10.0
idna==3.10
imageio==2.37.0
immutabledict==4.2.0
importlib_metadata==8.7.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.7
isoduration==20.11.0
jax==0.5.3
jax-ai-stack==2025.4.9
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.5.3
jaxtyping==0.2.34
jedi==0.19.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.5.0
kiwisolver==1.4.7
lark==1.2.2
libclang==18.1.1
lineax==0.0.5
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_collections==0.1.1
ml_dtypes==0.5.1
msgpack==1.1.0
mypy_extensions==1.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.4
optimistix==0.0.7
optree==0.12.1
orbax-checkpoint==0.11.12
orbax-export==0.0.6
overrides==7.7.0
packaging==24.1
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.4
prometheus_client==0.22.1
promise==2.3
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
python-box==7.3.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyzmq==26.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.8.1
rpds-py==0.26.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.14.1
Send2Trash==1.8.3
sentry-sdk==2.14.0
setproctitle==1.3.3
simple-parsing==0.1.6
simplejson==3.20.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.8
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
tensorstore==0.1.76
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.55.4
treescope==0.1.9
triton==3.1.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3==2.2.3
wandb==0.18.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.14
wrapt==1.16.0
zipp==3.20.2
